backend/requirement.txt
flask
flask-pymongo
flask-bcrypt
flask-cors
